requirements.txt
annotated-types==0.7.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
anyio==4.9.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
attrs==25.3.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
beautifulsoup4==4.13.4 ; python_version >= "3.11" and python_full_version <= "3.13.0"
cachetools==5.5.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
certifi==2025.1.31 ; python_version >= "3.11" and python_full_version <= "3.13.0"
cffi==1.17.1 ; python_version >= "3.11" and python_full_version <= "3.13.0" and os_name == "nt" and implementation_name != "pypy"
charset-normalizer==3.4.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
click==8.1.8 ; python_version >= "3.11" and python_full_version <= "3.13.0"
colorama==0.4.6 ; python_version >= "3.11" and python_full_version <= "3.13.0" and platform_system == "Windows"
db-dtypes==1.4.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
fastapi==0.115.12 ; python_version >= "3.11" and python_full_version <= "3.13.0"
ftfy==6.3.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
google-api-core==2.24.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
google-auth-oauthlib==1.2.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
google-auth==2.39.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
google-cloud-bigquery==3.31.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
google-cloud-core==2.4.3 ; python_version >= "3.11" and python_full_version <= "3.13.0"
google-cloud-secret-manager==2.23.3 ; python_version >= "3.11" and python_full_version <= "3.13.0"
google-cloud-storage==3.1.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
google-crc32c==1.7.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
google-resumable-media==2.7.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
googleapis-common-protos==1.70.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
grpc-google-iam-v1==0.14.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
grpcio-status==1.62.3 ; python_version >= "3.11" and python_full_version <= "3.13.0"
grpcio==1.71.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
h11==0.14.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
idna==3.10 ; python_version >= "3.11" and python_full_version <= "3.13.0"
numpy==1.26.4 ; python_version >= "3.11" and python_full_version <= "3.13.0"
oauthlib==3.2.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
outcome==1.3.0.post0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
packaging==24.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
pandas-gbq==0.25.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
pandas==2.2.3 ; python_version >= "3.11" and python_full_version <= "3.13.0"
proto-plus==1.26.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
protobuf==6.30.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
pyarrow==14.0.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
pyasn1-modules==0.4.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
pyasn1==0.6.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
pycparser==2.22 ; python_version >= "3.11" and python_full_version <= "3.13.0" and os_name == "nt" and implementation_name != "pypy"
pydantic-core==2.33.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
pydantic==2.11.4 ; python_version >= "3.11" and python_full_version <= "3.13.0"
pydata-google-auth==1.9.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
pysocks==1.7.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
python-dotenv==1.1.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
pytz==2025.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
requests-oauthlib==2.0.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
requests==2.32.3 ; python_version >= "3.11" and python_full_version <= "3.13.0"
rsa==4.9.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
selenium==4.31.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
setuptools==78.1.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
six==1.17.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
sortedcontainers==2.4.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
soupsieve==2.6 ; python_version >= "3.11" and python_full_version <= "3.13.0"
starlette==0.46.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
trio-websocket==0.12.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
trio==0.29.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
typing-extensions==4.13.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
typing-inspection==0.4.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
tzdata==2025.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
urllib3==2.4.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
uvicorn==0.34.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
wcwidth==0.2.13 ; python_version >= "3.11" and python_full_version <= "3.13.0"
webdriver-manager==4.0.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
websocket-client==1.8.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
wsproto==1.2.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
